requirements.txt
openai>=1.23.0
chromadb>=0.4.24
langchain>=0.2.0
python-dotenv>=1.0.1
tqdm
matplotlib
